requirements.txt
ImageHash==4.2.0
numpy==1.20.1
opencv-python==4.5.1.48
Pillow==8.1.0
PyWavelets==1.1.1
scipy==1.6.1
six==1.15.0
